Stay in cask repo when repairing
For a first-time user, this instruction leaves them in
`/usr/local/Homebrew` unexpectedly. Everything works fine... it's just
weird.

The only important think about being in `brew --repository` is that it
is known to be a gitlab repo, so it's a good, confident bet, fine to
keep using.

But by running this in a subshell, it will keep the user in their
existing directory